test(windows_foundation): refactor tests
