--- packages/windows_foundation/test/collections_test.dart
@@ -1,1116 +1,0 @@
-// Copyright (c) 2023, Dart | Windows. Please see the AUTHORS file for details.
-// All rights reserved. Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-@TestOn('windows')
-
-import 'dart:ffi';
-
-import 'package:ffi/ffi.dart';
-import 'package:test/test.dart';
-import 'package:win32/win32.dart';
-import 'package:windows_foundation/internal.dart';
-import 'package:windows_foundation/windows_foundation.dart';
-
-// Exhaustively test the WinRT collection interfaces to make sure constructors,
-// properties and methods are working correctly.
-
-void main() {
-  if (!isWindowsRuntimeAvailable()) {
-    print('Skipping tests because Windows Runtime is not available.');
-    return;
-  }
-
-  group('IMap<String, Object?> (PropertySet)', () {
-    IMap<String, Object?> getMap() {
-      final guid = Guid.parse(IID_ISpVoice);
-      return IMap.empty()
-        ..insert('key1', null)
-        ..insert('key2', StringMap())
-        ..insert('key3', true)
-        ..insert('key4', DateTime(2022, 7, 11, 17, 30))
-        ..insert('key5', 0.5)
-        ..insert('key6', const Duration(seconds: 30))
-        ..insert('key7', guid)
-        ..insert('key8', 259)
-        ..insert('key9', const Point(3, -3))
-        ..insert('key10', const Rect(2, -2, 200, 100))
-        ..insert('key11', const Size(300, 1500))
-        ..insert('key12', 'strVal')
-        ..insert('key13', [true, false])
-        ..insert('key14',
-            [DateTime(2020, 7, 11, 17, 30), DateTime(2022, 7, 11, 17, 30)])
-        ..insert('key15', [2.5, 0.99])
-        ..insert('key16', const [Duration(hours: 1), Duration(minutes: 60)])
-        ..insert('key17', [guid])
-        ..insert('key18', [StringMap()])
-        ..insert('key19', [2022, -2022])
-        ..insert('key20', [const Point(3, -3)])
-        ..insert('key21', [const Rect(2, -2, 200, 100)])
-        ..insert('key22', [const Size(300, 1500)])
-        ..insert('key23', ['str1', 'str2']);
-    }
-
-    test('fromMap', () {
-      final map = IMap<String, Object?>.fromMap(
-          {'key1': StringMap(), 'key2': 259, 'key3': 'strVal'});
-
-      final stringMapVal = map.lookup('key1');
-      expect(stringMapVal, isA<IInspectable>());
-      expect(getClassName(stringMapVal as IInspectable),
-          equals('Windows.Foundation.Collections.StringMap'));
-      expect(map.lookup('key2'), equals(259));
-      expect(map.lookup('key3'), equals('strVal'));
-    });
-
-    test('lookup throws a WindowsException if the map is empty', () {
-      final map = getMap()..clear();
-      expect(() => map.lookup('key1'), throwsA(isA<WindowsException>()));
-    });
-
-    test('lookup throws a WindowsException if the item does not exists', () {
-      final map = getMap();
-      expect(() => map.lookup('key0'), throwsA(isA<WindowsException>()));
-    });
-
-    test('lookup returns items', () {
-      final map = getMap();
-      expect(map.lookup('key1'), isNull);
-
-      final stringMapVal = map.lookup('key2');
-      expect(stringMapVal, isA<IInspectable>());
-      expect(getClassName(stringMapVal as IInspectable),
-          equals('Windows.Foundation.Collections.StringMap'));
-
-      expect(map.lookup('key3'), isTrue);
-
-      final dateTimeVal = map.lookup('key4');
-      expect(dateTimeVal, isA<DateTime>());
-      final dateTime = dateTimeVal as DateTime;
-      expect(dateTime.millisecondsSinceEpoch,
-          equals(DateTime(2022, 7, 11, 17, 30).millisecondsSinceEpoch));
-
-      expect(map.lookup('key5'), equals(0.5));
-      expect(map.lookup('key6'), equals(const Duration(seconds: 30)));
-
-      final guidVal = map.lookup('key7');
-      expect(guidVal, isA<Guid>());
-      final guid = guidVal as Guid;
-      expect(guid.toString(), equals(IID_ISpVoice));
-
-      expect(map.lookup('key8'), equals(259));
-
-      final pointVal = map.lookup('key9');
-      expect(pointVal, isA<Point>());
-      final point = pointVal as Point;
-      expect(point.x, equals(3));
-      expect(point.y, equals(-3));
-
-      final rectVal = map.lookup('key10');
-      expect(rectVal, isA<Rect>());
-      final rect = rectVal as Rect;
-      expect(rect.height, equals(100));
-      expect(rect.width, equals(200));
-      expect(rect.x, equals(2));
-      expect(rect.y, equals(-2));
-
-      final sizeVal = map.lookup('key11');
-      expect(sizeVal, isA<Size>());
-      final size = sizeVal as Size;
-      expect(size.height, equals(1500));
-      expect(size.width, equals(300));
-
-      expect(map.lookup('key12'), equals('strVal'));
-
-      expect(map.lookup('key13'), equals([true, false]));
-
-      final dateTimeListVal = map.lookup('key14');
-      expect(dateTimeListVal, isA<List<DateTime>>());
-      final dateTimeList = dateTimeListVal as List<DateTime>;
-      expect(dateTimeList.first.millisecondsSinceEpoch,
-          equals(DateTime(2020, 7, 11, 17, 30).millisecondsSinceEpoch));
-      expect(dateTimeList.last.millisecondsSinceEpoch,
-          equals(DateTime(2022, 7, 11, 17, 30).millisecondsSinceEpoch));
-
-      expect(map.lookup('key15'), equals([2.5, 0.99]));
-
-      expect(map.lookup('key16'),
-          equals(const [Duration(hours: 1), Duration(minutes: 60)]));
-
-      final guidListVal = map.lookup('key17');
-      expect(guidListVal, isA<List<Guid>>());
-      final guidList = guidListVal as List<Guid>;
-      expect(guidList.first.toString(), equals(IID_ISpVoice));
-
-      final stringMapListVal = map.lookup('key18');
-      expect(stringMapListVal, isA<List<Object?>>());
-      final stringMapList = stringMapListVal as List<Object?>;
-      expect(getClassName(stringMapList.first as IInspectable),
-          equals('Windows.Foundation.Collections.StringMap'));
-
-      expect(map.lookup('key19'), equals([2022, -2022]));
-
-      final pointListVal = map.lookup('key20');
-      expect(pointListVal, isA<List<Point>>());
-      final pointList = pointListVal as List<Point>;
-      expect(pointList.first.x, equals(3));
-      expect(pointList.first.y, equals(-3));
-
-      final rectListVal = map.lookup('key21');
-      expect(rectListVal, isA<List<Rect>>());
-      final rectList = rectListVal as List<Rect>;
-      expect(rectList.first.height, equals(100));
-      expect(rectList.first.width, equals(200));
-      expect(rectList.first.x, equals(2));
-      expect(rectList.first.y, equals(-2));
-
-      final sizeListVal = map.lookup('key22');
-      expect(sizeListVal, isA<List<Size>>());
-      final sizeList = sizeListVal as List<Size>;
-      expect(sizeList.first.height, equals(1500));
-      expect(sizeList.first.width, equals(300));
-
-      expect(map.lookup('key23'), equals(['str1', 'str2']));
-    });
-
-    test('hasKey finds items', () {
-      final map = getMap();
-      expect(map.hasKey('key1'), isTrue);
-      expect(map.hasKey('key11'), isTrue);
-      expect(map.hasKey('key23'), isTrue);
-    });
-
-    test('hasKey returns false if the item does not exists', () {
-      final map = getMap();
-      expect(map.hasKey('key0'), isFalse);
-    });
-
-    test('getView', () {
-      final map = getMap();
-      final unmodifiableMap = map.getView();
-      expect(unmodifiableMap.length, equals(23));
-      expect(() => unmodifiableMap..clear(), throwsUnsupportedError);
-    });
-
-    test('insert replaces an existing item', () {
-      final map = getMap();
-      expect(map.size, equals(23));
-      expect(map.insert('key12', 'strValNew'), isTrue);
-      expect(map.size, equals(23));
-      expect(map.lookup('key12'), equals('strValNew'));
-    });
-
-    test('insert inserts a new item', () {
-      final map = getMap();
-      expect(map.size, equals(23));
-      expect(map.insert('key24', null), isFalse);
-      expect(map.size, equals(24));
-      expect(map.lookup('key24'), isNull);
-    });
-
-    test('remove throws a WindowsException if the map is empty', () {
-      final map = getMap()..clear();
-      expect(() => map.remove('key1'), throwsA(isA<WindowsException>()));
-    });
-
-    test('remove throws a WindowsException if the item does not exists', () {
-      final map = getMap();
-      expect(() => map.remove('key0'), throwsA(isA<WindowsException>()));
-    });
-
-    test('remove', () {
-      final map = getMap();
-      expect(map.size, equals(23));
-
-      map.remove('key1');
-      expect(map.size, equals(22));
-      expect(() => map.lookup('key1'), throwsA(isA<WindowsException>()));
-
-      map.remove('key6');
-      expect(map.size, equals(21));
-      expect(() => map.lookup('key6'), throwsA(isA<WindowsException>()));
-    });
-
-    test('clear', () {
-      final map = getMap();
-      expect(map.size, equals(23));
-      map.clear();
-      expect(map.size, equals(0));
-    });
-
-    test('toMap', () {
-      final map = getMap();
-      final dartMap = map.toMap();
-      expect(dartMap.length, equals(23));
-      expect(dartMap['key8'], equals(259));
-      expect(dartMap['key12'], equals('strVal'));
-      expect(dartMap['key15'], equals([2.5, 0.99]));
-      expect(() => dartMap..clear(), throwsUnsupportedError);
-    });
-
-    test('first', () {
-      final map =
-          IMap<String, Object?>.fromMap({'key1': 'icalendar', 'key2': 259});
-
-      final iterator = map.first();
-      expect(iterator.hasCurrent, isTrue);
-      var current = iterator.current;
-      expect(current.key, equals('key2'));
-      expect(current.value, equals(259));
-      expect(iterator.moveNext(), isTrue);
-      current = iterator.current;
-      expect(current.key, equals('key1'));
-      expect(current.value, equals('icalendar'));
-      expect(iterator.moveNext(), isFalse);
-    });
-
-    test('operator []', () {
-      final map = getMap();
-      expect(map['key1'], isNull);
-      expect(map['key3'], isTrue);
-      expect(map['key5'], equals(0.5));
-    });
-
-    test('operator []=', () {
-      final map = getMap();
-
-      // Replace an existing item.
-      expect(map.size, equals(23));
-      map['key5'] = null;
-      expect(map.size, equals(23));
-      expect(map.lookup('key5'), isNull);
-
-      // Insert a new item.
-      map['key24'] = 24;
-      expect(map.size, equals(24));
-      expect(map.lookup('key24'), equals(24));
-    });
-  });
-
-  group('IMap<String, Object?> (ValueSet)', () {
-    IMap<String, Object?> getMap() {
-      final guid = Guid.parse(IID_ISpVoice);
-      final valueSet = ValueSet()
-        ..insert('key1', null)
-        ..insert('key2', 'strVal');
-      return ValueSet()
-        ..insert('key1', null)
-        ..insert('key2', valueSet)
-        ..insert('key3', true)
-        ..insert('key4', DateTime(2022, 7, 11, 17, 30))
-        ..insert('key5', 0.5)
-        ..insert('key6', const Duration(seconds: 30))
-        ..insert('key7', guid)
-        ..insert('key8', 259)
-        ..insert('key9', const Point(3, -3))
-        ..insert('key10', const Rect(2, -2, 200, 100))
-        ..insert('key11', const Size(300, 1500))
-        ..insert('key12', 'strVal')
-        ..insert('key13', [true, false])
-        ..insert('key14',
-            [DateTime(2020, 7, 11, 17, 30), DateTime(2022, 7, 11, 17, 30)])
-        ..insert('key15', [2.5, 0.99])
-        ..insert('key16', const [Duration(hours: 1), Duration(minutes: 60)])
-        ..insert('key17', [guid])
-        ..insert('key18', [2022, -2022])
-        ..insert('key19', [const Point(3, -3)])
-        ..insert('key20', [const Rect(2, -2, 200, 100)])
-        ..insert('key21', [const Size(300, 1500)])
-        ..insert('key22', ['str1', 'str2']);
-    }
-
-    test('lookup throws a WindowsException if the map is empty', () {
-      final map = getMap()..clear();
-      expect(() => map.lookup('key1'), throwsA(isA<WindowsException>()));
-    });
-
-    test('lookup throws a WindowsException if the item does not exists', () {
-      final map = getMap();
-      expect(() => map.lookup('key0'), throwsA(isA<WindowsException>()));
-    });
-
-    test('lookup returns items', () {
-      final map = getMap();
-      expect(map.lookup('key1'), isNull);
-
-      final valueSetVal = map.lookup('key2');
-      expect(valueSetVal, isA<IInspectable>());
-      final valueSet = ValueSet.fromPtr(
-          (valueSetVal as IInspectable).toInterface(IID_IMap_String_Object));
-      expect(getClassName(valueSet),
-          equals('Windows.Foundation.Collections.ValueSet'));
-      expect(valueSet.size, equals(2));
-      expect(valueSet.lookup('key1'), isNull);
-      expect(valueSet.lookup('key2'), equals('strVal'));
-
-      expect(map.lookup('key3'), isTrue);
-
-      final dateTimeVal = map.lookup('key4');
-      expect(dateTimeVal, isA<DateTime>());
-      final dateTime = dateTimeVal as DateTime;
-      expect(dateTime.millisecondsSinceEpoch,
-          equals(DateTime(2022, 7, 11, 17, 30).millisecondsSinceEpoch));
-
-      expect(map.lookup('key5'), equals(0.5));
-      expect(map.lookup('key6'), equals(const Duration(seconds: 30)));
-
-      final guidVal = map.lookup('key7');
-      expect(guidVal, isA<Guid>());
-      final guid = guidVal as Guid;
-      expect(guid.toString(), equals(IID_ISpVoice));
-
-      expect(map.lookup('key8'), equals(259));
-
-      final pointVal = map.lookup('key9');
-      expect(pointVal, isA<Point>());
-      final point = pointVal as Point;
-      expect(point.x, equals(3));
-      expect(point.y, equals(-3));
-
-      final rectVal = map.lookup('key10');
-      expect(rectVal, isA<Rect>());
-      final rect = rectVal as Rect;
-      expect(rect.height, equals(100));
-      expect(rect.width, equals(200));
-      expect(rect.x, equals(2));
-      expect(rect.y, equals(-2));
-
-      final sizeVal = map.lookup('key11');
-      expect(sizeVal, isA<Size>());
-      final size = sizeVal as Size;
-      expect(size.height, equals(1500));
-      expect(size.width, equals(300));
-
-      expect(map.lookup('key12'), equals('strVal'));
-
-      expect(map.lookup('key13'), equals([true, false]));
-
-      final dateTimeListVal = map.lookup('key14');
-      expect(dateTimeListVal, isA<List<DateTime>>());
-      final dateTimeList = dateTimeListVal as List<DateTime>;
-      expect(dateTimeList.first.millisecondsSinceEpoch,
-          equals(DateTime(2020, 7, 11, 17, 30).millisecondsSinceEpoch));
-      expect(dateTimeList.last.millisecondsSinceEpoch,
-          equals(DateTime(2022, 7, 11, 17, 30).millisecondsSinceEpoch));
-
-      expect(map.lookup('key15'), equals([2.5, 0.99]));
-
-      expect(map.lookup('key16'),
-          equals(const [Duration(hours: 1), Duration(minutes: 60)]));
-
-      final guidListVal = map.lookup('key17');
-      expect(guidListVal, isA<List<Guid>>());
-      final guidList = guidListVal as List<Guid>;
-      expect(guidList.first.toString(), equals(IID_ISpVoice));
-
-      expect(map.lookup('key18'), equals([2022, -2022]));
-
-      final pointListVal = map.lookup('key19');
-      expect(pointListVal, isA<List<Point>>());
-      final pointList = pointListVal as List<Point>;
-      expect(pointList.first.x, equals(3));
-      expect(pointList.first.y, equals(-3));
-
-      final rectListVal = map.lookup('key20');
-      expect(rectListVal, isA<List<Rect>>());
-      final rectList = rectListVal as List<Rect>;
-      expect(rectList.first.height, equals(100));
-      expect(rectList.first.width, equals(200));
-      expect(rectList.first.x, equals(2));
-      expect(rectList.first.y, equals(-2));
-
-      final sizeListVal = map.lookup('key21');
-      expect(sizeListVal, isA<List<Size>>());
-      final sizeList = sizeListVal as List<Size>;
-      expect(sizeList.first.height, equals(1500));
-      expect(sizeList.first.width, equals(300));
-
-      expect(map.lookup('key22'), equals(['str1', 'str2']));
-    });
-
-    test('hasKey finds items', () {
-      final map = getMap();
-      expect(map.hasKey('key1'), isTrue);
-      expect(map.hasKey('key11'), isTrue);
-      expect(map.hasKey('key22'), isTrue);
-    });
-
-    test('hasKey returns false if the item does not exists', () {
-      final map = getMap();
-      expect(map.hasKey('key0'), isFalse);
-    });
-
-    test('getView', () {
-      final map = getMap();
-      final unmodifiableMap = map.getView();
-      expect(unmodifiableMap.length, equals(22));
-      expect(() => unmodifiableMap..clear(), throwsUnsupportedError);
-    });
-
-    test('insert replaces an existing item', () {
-      final map = getMap();
-      expect(map.size, equals(22));
-      expect(map.insert('key12', 'strValNew'), isTrue);
-      expect(map.size, equals(22));
-      expect(map.lookup('key12'), equals('strValNew'));
-    });
-
-    test('insert inserts a new item', () {
-      final map = getMap();
-      expect(map.size, equals(22));
-      expect(map.insert('key23', null), isFalse);
-      expect(map.size, equals(23));
-      expect(map.lookup('key23'), isNull);
-    });
-
-    test('remove throws a WindowsException if the map is empty', () {
-      final map = getMap()..clear();
-      expect(() => map.remove('key1'), throwsA(isA<WindowsException>()));
-    });
-
-    test('remove throws a WindowsException if the item does not exists', () {
-      final map = getMap();
-      expect(() => map.remove('key0'), throwsA(isA<WindowsException>()));
-    });
-
-    test('remove', () {
-      final map = getMap();
-      expect(map.size, equals(22));
-
-      map.remove('key1');
-      expect(map.size, equals(21));
-      expect(() => map.lookup('key1'), throwsA(isA<WindowsException>()));
-
-      map.remove('key6');
-      expect(map.size, equals(20));
-      expect(() => map.lookup('key6'), throwsA(isA<WindowsException>()));
-    });
-
-    test('clear', () {
-      final map = getMap();
-      expect(map.size, equals(22));
-      map.clear();
-      expect(map.size, equals(0));
-    });
-
-    test('toMap', () {
-      final map = getMap();
-      final dartMap = map.toMap();
-      expect(dartMap.length, equals(22));
-      expect(dartMap['key8'], equals(259));
-      expect(dartMap['key12'], equals('strVal'));
-      expect(dartMap['key15'], equals([2.5, 0.99]));
-      expect(() => dartMap..clear(), throwsUnsupportedError);
-    });
-
-    test('first', () {
-      final map = ValueSet()
-        ..insert('key1', 'icalendar')
-        ..insert('key2', 259);
-
-      final iterator = map.first();
-      expect(iterator.hasCurrent, isTrue);
-      expect(iterator.current.key, equals('key2'));
-      expect(iterator.current.value, equals(259));
-      expect(iterator.moveNext(), isTrue);
-      expect(iterator.current.key, equals('key1'));
-      expect(iterator.current.value, equals('icalendar'));
-      expect(iterator.moveNext(), isFalse);
-    });
-
-    test('operator []', () {
-      final map = getMap();
-      expect(map['key1'], isNull);
-      expect(map['key3'], isTrue);
-      expect(map['key5'], equals(0.5));
-    });
-
-    test('operator []=', () {
-      final map = getMap();
-
-      // Replace an existing item.
-      expect(map.size, equals(22));
-      map['key5'] = null;
-      expect(map.size, equals(22));
-      expect(map.lookup('key5'), isNull);
-
-      // Insert a new item.
-      map['key23'] = 23;
-      expect(map.size, equals(23));
-      expect(map.lookup('key23'), equals(23));
-    });
-  });
-
-  group('IMap<String, String> (StringMap)', () {
-    IMap<String, String> getMap() => IMap.empty()
-      ..insert('key1', 'value1')
-      ..insert('key2', '')
-      ..insert('key3', 'value3');
-
-    test('fromMap', () {
-      final map =
-          IMap.fromMap({'key1': 'value1', 'key2': '', 'key3': 'value3'});
-      expect(map.lookup('key1'), equals('value1'));
-      expect(map.lookup('key2'), isEmpty);
-      expect(map.lookup('key3'), equals('value3'));
-    });
-
-    test('lookup throws a WindowsException if the map is empty', () {
-      final map = getMap()..clear();
-      expect(() => map.lookup('key1'), throwsA(isA<WindowsException>()));
-    });
-
-    test('lookup throws a WindowsException if the item does not exists', () {
-      final map = getMap();
-      expect(() => map.lookup('key4'), throwsA(isA<WindowsException>()));
-    });
-
-    test('lookup returns items', () {
-      final map = getMap();
-      expect(map.lookup('key1'), equals('value1'));
-      expect(map.lookup('key2'), isEmpty);
-      expect(map.lookup('key3'), equals('value3'));
-    });
-
-    test('hasKey finds items', () {
-      final map = getMap();
-      expect(map.hasKey('key1'), isTrue);
-      expect(map.hasKey('key2'), isTrue);
-      expect(map.hasKey('key3'), isTrue);
-    });
-
-    test('hasKey returns false if the item does not exists', () {
-      final map = getMap();
-      expect(map.hasKey('key4'), isFalse);
-    });
-
-    test('getView', () {
-      final map = getMap();
-      final unmodifiableMap = map.getView();
-      expect(unmodifiableMap.length, equals(3));
-      expect(() => unmodifiableMap..clear(), throwsUnsupportedError);
-    });
-
-    test('insert replaces an existing item', () {
-      final map = getMap();
-      expect(map.size, equals(3));
-      expect(map.insert('key1', 'value1New'), isTrue);
-      expect(map.size, equals(3));
-      expect(map.lookup('key1'), equals('value1New'));
-    });
-
-    test('insert inserts a new item', () {
-      final map = getMap();
-      expect(map.size, equals(3));
-      expect(map.insert('key4', 'value4'), isFalse);
-      expect(map.size, equals(4));
-      expect(map.lookup('key4'), equals('value4'));
-    });
-
-    test('remove throws a WindowsException if the map is empty', () {
-      final map = getMap()..clear();
-      expect(() => map.remove('key0'), throwsA(isA<WindowsException>()));
-    });
-
-    test('remove throws a WindowsException if the item does not exists', () {
-      final map = getMap();
-      expect(() => map.remove('key4'), throwsA(isA<WindowsException>()));
-    });
-
-    test('remove', () {
-      final map = getMap();
-      expect(map.size, equals(3));
-      map.remove('key1');
-      expect(map.size, equals(2));
-      expect(() => map.lookup('key1'), throwsA(isA<WindowsException>()));
-
-      map.remove('key2');
-      expect(map.size, equals(1));
-      expect(() => map.lookup('key2'), throwsA(isA<WindowsException>()));
-    });
-
-    test('clear', () {
-      final map = getMap();
-      expect(map.size, equals(3));
-      map.clear();
-      expect(map.size, equals(0));
-    });
-
-    test('toMap', () {
-      final map = getMap();
-      final dartMap = map.toMap();
-      expect(dartMap.length, equals(3));
-      expect(dartMap['key1'], equals('value1'));
-      expect(dartMap['key2'], isEmpty);
-      expect(dartMap['key3'], equals('value3'));
-      expect(() => dartMap..clear(), throwsUnsupportedError);
-    });
-
-    test('first', () {
-      final map = getMap();
-      final iterator = map.first();
-      expect(iterator.hasCurrent, isTrue);
-      var current = iterator.current;
-      expect(current.key, equals('key3'));
-      expect(current.value, equals('value3'));
-      expect(iterator.moveNext(), isTrue);
-      current = iterator.current;
-      expect(current.key, equals('key2'));
-      expect(current.value, isEmpty);
-      expect(iterator.moveNext(), isTrue);
-      current = iterator.current;
-      expect(current.key, equals('key1'));
-      expect(current.value, equals('value1'));
-      expect(iterator.moveNext(), isFalse);
-    });
-
-    test('operator []', () {
-      final map = getMap();
-      expect(map['key1'], equals('value1'));
-      expect(map['key2'], isEmpty);
-      expect(map['key3'], equals('value3'));
-    });
-
-    test('operator []=', () {
-      final map = getMap();
-
-      // Replace an existing item.
-      expect(map.size, equals(3));
-      map['key3'] = 'newValue3';
-      expect(map.size, equals(3));
-      expect(map.lookup('key3'), equals('newValue3'));
-
-      // Insert a new item.
-      map['key4'] = 'value4';
-      expect(map.size, equals(4));
-      expect(map.lookup('key4'), equals('value4'));
-    });
-  });
-
-  group('IMapView<String, String> (StringMap)', () {
-    IMapView<String, String> getView(Pointer<COMObject> ptr) {
-      final retValuePtr = calloc<COMObject>();
-
-      final hr = ptr.ref.lpVtbl.value
-              .elementAt(9)
-              .cast<
-                  Pointer<
-                      NativeFunction<
-                          HRESULT Function(VTablePointer lpVtbl,
-                              Pointer<COMObject> retValuePtr)>>>()
-              .value
-              .asFunction<
-                  int Function(
-                      VTablePointer lpVtbl, Pointer<COMObject> retValuePtr)>()(
-          ptr.ref.lpVtbl, retValuePtr);
-
-      if (FAILED(hr)) throwWindowsException(hr);
-
-      return IMapView.fromPtr(retValuePtr,
-          iterableIid: IID_IIterable_IKeyValuePair_String_String);
-    }
-
-    IMapView<String, String> getMapView() {
-      final map = IMap<String, String>.empty()
-        ..insert('key1', 'value1')
-        ..insert('key2', '')
-        ..insert('key3', 'value3');
-      return getView(map.ptr);
-    }
-
-    test('lookup throws a WindowsException if the map is empty', () {
-      final map = IMap<String, String>.empty();
-      final mapView = getView(map.ptr);
-      expect(() => mapView.lookup('key1'), throwsA(isA<WindowsException>()));
-    });
-
-    test('lookup throws a WindowsException if the item does not exists', () {
-      final mapView = getMapView();
-      expect(() => mapView.lookup('key4'), throwsA(isA<WindowsException>()));
-    });
-
-    test('lookup returns items', () {
-      final mapView = getMapView();
-      expect(mapView.lookup('key1'), equals('value1'));
-      expect(mapView.lookup('key2'), isEmpty);
-      expect(mapView.lookup('key3'), equals('value3'));
-    });
-
-    test('hasKey finds items', () {
-      final mapView = getMapView();
-      expect(mapView.hasKey('key1'), isTrue);
-      expect(mapView.hasKey('key2'), isTrue);
-      expect(mapView.hasKey('key3'), isTrue);
-    });
-
-    test('hasKey returns false if the item does not exists', () {
-      final mapView = getMapView();
-      expect(mapView.hasKey('key4'), isFalse);
-    });
-
-    test('toMap', () {
-      final mapView = getMapView();
-      final dartMap = mapView.toMap();
-      expect(dartMap.length, equals(3));
-      expect(dartMap['key1'], equals('value1'));
-      expect(dartMap['key2'], isEmpty);
-      expect(dartMap['key3'], equals('value3'));
-      expect(() => dartMap..clear(), throwsUnsupportedError);
-    });
-
-    test('first', () {
-      final mapView = getMapView();
-      final iterator = mapView.first();
-      expect(iterator.hasCurrent, isTrue);
-      var current = iterator.current;
-      expect(current.key, equals('key3'));
-      expect(current.value, equals('value3'));
-      expect(iterator.moveNext(), isTrue);
-      current = iterator.current;
-      expect(current.key, equals('key2'));
-      expect(current.value, isEmpty);
-      expect(iterator.moveNext(), isTrue);
-      current = iterator.current;
-      expect(current.key, equals('key1'));
-      expect(current.value, equals('value1'));
-      expect(iterator.moveNext(), isFalse);
-    });
-
-    test('operator []', () {
-      final map = getMapView();
-      expect(map['key1'], equals('value1'));
-      expect(map['key2'], isEmpty);
-      expect(map['key3'], equals('value3'));
-    });
-  });
-
-  group('IVector<Uri?>', () {
-    IVector<Uri?> getServerUris(Pointer<COMObject> ptr) {
-      final retValuePtr = calloc<COMObject>();
-
-      final hr = ptr.ref.vtable
-              .elementAt(6)
-              .cast<
-                  Pointer<
-                      NativeFunction<
-                          HRESULT Function(VTablePointer lpVtbl,
-                              Pointer<COMObject> retValuePtr)>>>()
-              .value
-              .asFunction<
-                  int Function(
-                      VTablePointer lpVtbl, Pointer<COMObject> retValuePtr)>()(
-          ptr.ref.lpVtbl, retValuePtr);
-
-      if (FAILED(hr)) {
-        free(retValuePtr);
-        throwWindowsException(hr);
-      }
-
-      return IVector.fromPtr(retValuePtr, iterableIid: IID_IIterable_Uri);
-    }
-
-    IVector<Uri?> getVector() {
-      // ignore: constant_identifier_names
-      const IID_IVpnPlugInProfile = '{0edf0da4-4f00-4589-8d7b-4bf988f6542c}';
-      final object = createObject(IInspectable.new,
-          'Windows.Networking.Vpn.VpnPlugInProfile', IID_IVpnPlugInProfile);
-      return getServerUris(object.ptr);
-    }
-
-    test('getAt throws a WindowsException if the vector is empty', () {
-      final vector = getVector();
-      expect(() => vector.getAt(0), throwsA(isA<WindowsException>()));
-    });
-
-    test('getAt throws a WindowsException if the index is out of bounds', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(Uri.parse('https://dart.dev/docs'));
-      expect(() => vector.getAt(2), throwsA(isA<WindowsException>()));
-    });
-
-    test('getAt returns elements', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(null)
-        ..append(Uri.parse('https://dart.dev/docs'));
-      expect(vector.size, equals(3));
-      expect(vector.getAt(0), equals(Uri.parse('https://dart.dev/overview')));
-      expect(vector.getAt(1), isNull);
-      expect(vector.getAt(2), equals(Uri.parse('https://dart.dev/docs')));
-    });
-
-    test('getView returns empty List', () {
-      final vector = getVector();
-      final list = vector.getView();
-      expect(list, isEmpty);
-    });
-
-    test('getView returns elements', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(null)
-        ..append(Uri.parse('https://dart.dev/docs'));
-      final list = vector.getView();
-      expect(list.length, equals(3));
-      expect(() => list..clear(), throwsUnsupportedError);
-    });
-
-    test('setAt throws a WindowsException if the vector is empty', () {
-      final vector = getVector();
-      expect(() => vector.setAt(0, Uri.parse('https://dart.dev/overview')),
-          throwsA(isA<WindowsException>()));
-    });
-
-    test('setAt throws a WindowsException if the index is out of bounds', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(Uri.parse('https://dart.dev/docs'));
-      expect(
-          () => vector.setAt(3, Uri.parse('https://flutter.dev/development')),
-          throwsA(isA<WindowsException>()));
-    });
-
-    test('setAt', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(null);
-      expect(vector.size, equals(2));
-      vector.setAt(0, Uri.parse('https://flutter.dev/development'));
-      expect(vector.size, equals(2));
-      vector.setAt(1, Uri.parse('https://flutter.dev/multi-platform'));
-      expect(vector.size, equals(2));
-      expect(vector.getAt(0),
-          equals(Uri.parse('https://flutter.dev/development')));
-      expect(vector.getAt(1),
-          equals(Uri.parse('https://flutter.dev/multi-platform')));
-    });
-
-    test('insertAt throws a WindowsException if the index is out of bounds',
-        () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(Uri.parse('https://dart.dev/docs'));
-      expect(
-          () =>
-              vector.insertAt(3, Uri.parse('https://flutter.dev/development')),
-          throwsA(isA<WindowsException>()));
-    });
-
-    test('insertAt', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(Uri.parse('https://dart.dev/docs'));
-      expect(vector.size, equals(2));
-      vector.insertAt(0, Uri.parse('https://flutter.dev/development'));
-      expect(vector.size, equals(3));
-      vector.insertAt(2, null);
-      expect(vector.size, equals(4));
-      expect(vector.getAt(0),
-          equals(Uri.parse('https://flutter.dev/development')));
-      expect(vector.getAt(1), equals(Uri.parse('https://dart.dev/overview')));
-      expect(vector.getAt(2), isNull);
-      expect(vector.getAt(3), equals(Uri.parse('https://dart.dev/docs')));
-    });
-
-    test('removeAt throws a WindowsException if the vector is empty', () {
-      final vector = getVector();
-      expect(() => vector.removeAt(0), throwsA(isA<WindowsException>()));
-    });
-
-    test('removeAt throws a WindowsException if the index is out of bounds',
-        () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(Uri.parse('https://dart.dev/docs'));
-      expect(() => vector.removeAt(3), throwsA(isA<WindowsException>()));
-    });
-
-    test('removeAt', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(Uri.parse('https://dart.dev/docs'))
-        ..append(null)
-        ..append(Uri.parse('https://flutter.dev/multi-platform'));
-      expect(vector.size, equals(4));
-      vector.removeAt(2);
-      expect(vector.size, equals(3));
-      expect(vector.getAt(2),
-          equals(Uri.parse('https://flutter.dev/multi-platform')));
-      vector.removeAt(0);
-      expect(vector.size, equals(2));
-      expect(vector.getAt(0), equals(Uri.parse('https://dart.dev/docs')));
-      vector.removeAt(1);
-      expect(vector.size, equals(1));
-      expect(vector.getAt(0), equals(Uri.parse('https://dart.dev/docs')));
-      vector.removeAt(0);
-      expect(vector.size, equals(0));
-    });
-
-    test('append', () {
-      final vector = getVector();
-      expect(vector.size, equals(0));
-      vector.append(Uri.parse('https://dart.dev/overview'));
-      expect(vector.size, equals(1));
-      vector.append(null);
-      expect(vector.size, equals(2));
-      vector.append(Uri.parse('https://dart.dev/docs'));
-      expect(vector.size, equals(3));
-    });
-
-    test('removeAtEnd throws a WindowsException if the vector is empty', () {
-      final vector = getVector();
-      expect(vector.removeAtEnd, throwsA(isA<WindowsException>()));
-    });
-
-    test('removeAtEnd', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(Uri.parse('https://dart.dev/docs'));
-      expect(vector.size, equals(2));
-      vector.removeAtEnd();
-      expect(vector.size, equals(1));
-    });
-
-    test('clear', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(null)
-        ..append(Uri.parse('https://dart.dev/docs'));
-      expect(vector.size, equals(3));
-      vector.clear();
-      expect(vector.size, equals(0));
-    });
-
-    test('getMany returns 0 if the vector is empty', () {
-      final vector = getVector();
-      final (itemCount, :items) = vector.getMany(0, 1);
-      expect(itemCount, equals(0));
-      expect(items.length, equals(0));
-    });
-
-    test('getMany returns elements starting from index 0', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(null)
-        ..append(Uri.parse('https://flutter.dev/development'));
-      final (itemCount, :items) = vector.getMany(0, 3);
-      expect(itemCount, equals(3));
-      expect(items.length, equals(3));
-      expect(items[0].toString(), equals('https://dart.dev/overview'));
-      expect(items[1], isNull);
-      expect(items[2].toString(), equals('https://flutter.dev/development'));
-    });
-
-    test(
-        'getMany returns all elements if valueSize is greater than the number '
-        'of elements', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(null)
-        ..append(Uri.parse('https://flutter.dev/development'));
-      final (itemCount, :items) = vector.getMany(0, 5);
-      expect(itemCount, equals(3));
-      expect(items.length, equals(3));
-      expect(items[0].toString(), equals('https://dart.dev/overview'));
-      expect(items[1], isNull);
-      expect(items[2].toString(), equals('https://flutter.dev/development'));
-    });
-
-    test('getMany returns elements starting from index 1', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(null)
-        ..append(Uri.parse('https://flutter.dev/development'));
-      final (itemCount, :items) = vector.getMany(1, 2);
-      expect(itemCount, equals(2));
-      expect(items.length, equals(2));
-      expect(items[0], isNull);
-      expect(items[1].toString(), equals('https://flutter.dev/development'));
-    });
-
-    test('replaceAll', () {
-      final vector = getVector();
-      expect(vector.size, equals(0));
-      vector.replaceAll([
-        Uri.parse('https://dart.dev/overview'),
-        null,
-        Uri.parse('https://dart.dev/docs')
-      ]);
-      expect(vector.size, equals(3));
-      expect(vector.getAt(0), equals(Uri.parse('https://dart.dev/overview')));
-      expect(vector.getAt(1), isNull);
-      expect(vector.getAt(2), equals(Uri.parse('https://dart.dev/docs')));
-      vector.replaceAll([
-        Uri.parse('https://flutter.dev/development'),
-        Uri.parse('https://flutter.dev/multi-platform')
-      ]);
-      expect(vector.size, equals(2));
-      expect(vector.getAt(0),
-          equals(Uri.parse('https://flutter.dev/development')));
-      expect(vector.getAt(1),
-          equals(Uri.parse('https://flutter.dev/multi-platform')));
-    });
-
-    test('toList', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(null)
-        ..append(Uri.parse('https://flutter.dev/development'));
-      final list = vector.toList();
-      expect(list.length, equals(3));
-      expect(list[0], equals(Uri.parse('https://dart.dev/overview')));
-      expect(list[1], isNull);
-      expect(list[2], equals(Uri.parse('https://flutter.dev/development')));
-      expect(() => list..clear(), throwsUnsupportedError);
-    });
-
-    test('first', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(null)
-        ..append(Uri.parse('https://flutter.dev/development'));
-      final iterator = vector.first();
-      expect(iterator.hasCurrent, isTrue);
-      expect(iterator.current, equals(Uri.parse('https://dart.dev/overview')));
-      expect(iterator.moveNext(), isTrue);
-      expect(iterator.current, isNull);
-      expect(iterator.moveNext(), isTrue);
-      expect(iterator.current,
-          equals(Uri.parse('https://flutter.dev/development')));
-      expect(iterator.moveNext(), isFalse);
-    });
-
-    test('operator []', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(null)
-        ..append(Uri.parse('https://flutter.dev/development'));
-      expect(vector[0], equals(Uri.parse('https://dart.dev/overview')));
-      expect(vector[1], isNull);
-      expect(vector[2], equals(Uri.parse('https://flutter.dev/development')));
-    });
-
-    test('operator []=', () {
-      final vector = getVector()
-        ..append(Uri.parse('https://dart.dev/overview'))
-        ..append(Uri.parse('https://dart.dev/docs'))
-        ..append(Uri.parse('https://flutter.dev/development'));
-      vector[1] = Uri.parse('https://dart.dev/tutorials');
-      expect(vector[1], equals(Uri.parse('https://dart.dev/tutorials')));
-    });
-
-    test('operator +', () {
-      final vector = getVector()..append(Uri.parse('https://dart.dev/docs'));
-      final list = [Uri.parse('https://dart.dev/tutorials')];
-      final newList = vector + list;
-      expect(newList.length, equals(2));
-      expect(
-          newList,
-          orderedEquals([
-            Uri.parse('https://dart.dev/docs'),
-            Uri.parse('https://dart.dev/tutorials')
-          ]));
-    });
-  });
-}
